requirements.txt
Flask==3.0.2
line-bot-sdk
openai==1.13.3
Gunicorn
uvicorn
gspread
oauth2client
httpx==0.27.2